fix: Change octaves from i32 to f32 to prevent infinite loop
The octaves field was declared as i32 in the uniform struct, but written as
a float using Float32Array. When WGSL read the float bytes as an integer,
it resulted in a huge number (1084227584 instead of 5), causing an infinite
loop in the FBM function and GPU hang.

Changed octaves to f32 in uniform struct and cast to i32 when calling fbm().

+++ src/codegen/compilers/ComputeNodes.js
@@ -144,7 +144,7 @@ struct Uniforms {
   resolution: vec2<f32>,
   time: f32,
   scale: f32,
-  octaves: i32,
+  octaves: f32,
   speed: f32,
   padding: vec2<f32>
 }
@@ -204,7 +204,7 @@ fn main(@builtin(global_invocation_id) global_id: vec3<u32>) {
   var noisePos = uv * uniforms.scale;
   noisePos += vec2<f32>(time * 0.1, time * 0.15);
 
-  let noiseValue = fbm(noisePos, uniforms.octaves);
+  let noiseValue = fbm(noisePos, i32(uniforms.octaves));
 
   ${colorize ? `
   // Colorize the noise
